src/main/java/usantatecla/movies/v21/Customer.java
package usantatecla.movies.v21;

import java.util.ArrayList;
import java.util.Iterator;
import java.util.List;

public class Customer {

	private String name;

	private List<Rental> rentals;

	public Customer(String name) {
		this.name = name;
		rentals = new ArrayList<Rental>();
	}

	public void addRental(Rental rental) {
		rentals.add(rental);
	}

	public String getName() {
		return name;
	}

	public String statement() {
		Iterator<Rental> rentals = this.rentals.iterator();
		String result = "Rental Record for " + this.getName() + "\n";
		while (rentals.hasNext()) {
			Rental rental = rentals.next();
			result += "\t" + rental.getTitle() + "\t" + String.valueOf(rental.getCharge()) + "\n";
		}
		result += "Amount owed is " + String.valueOf(this.getTotalCharge()) + "\n";
		result += "You earned " + String.valueOf(this.getTotalFrequentRenterPoints()) + " frequent renter points";
		return result;
	}

	private double getTotalCharge() {
		return this.rentals.stream()
				.mapToDouble(rental -> rental.getCharge())
				.sum();
	}

	private int getTotalFrequentRenterPoints() {
		return this.rentals.stream()
				.mapToInt(rental -> rental.getFrequentRenterPoints())
				.sum();
	}

}
